+++ react/components/EXPERIMENTAL_Select/styles.js
@@ -2,7 +2,6 @@ export const getFontClassNameFromSize = size => {
   switch (size) {
     case 'small':
       return 't-small'
-    case 'large':
     default:
       return 't-body'
   }