+++ react/components/Menu/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { Overlay } from 'react-overlays'
+import classNames from 'classnames'
 
 import Toggle from '../Toggle'
 import { withForwardedRef, refShape } from '../../modules/withForwardedRef'
@@ -174,8 +175,15 @@ class Menu extends Component {
                     className={menuHeight ? 'overflow-auto' : ''}>
                     {options.map((option, index) => (
                       <button
+                        disabled={option.disabled}
                         key={index}
-                        className="flex justify-between items-center t-body ph6 h-regular pointer hover-bg-muted-5 ma0 bg-transparent bn w-100 tl"
+                        className={classNames(
+                          'flex justify-between items-center t-body ph6 h-regular ma0 bg-transparent bn w-100 tl',
+                          {
+                            'hover-bg-muted-5 pointer': !option.disabled,
+                          }
+                        )}
+                        data-testid={option.testId || `menu-option-${index}`}
                         onClick={() => {
                           option.onClick(option)
                           if (onClose) {
@@ -183,9 +191,12 @@ class Menu extends Component {
                           }
                         }}>
                         <span
-                          className={`${
-                            option.toggle ? 'w-70 truncate' : 'w-100 truncate'
-                          } ${option.isDangerous ? 'c-danger' : ''}`}>
+                          className={classNames({
+                            'w-70 truncate': option.toggle,
+                            'w-100 truncate': !option.toggle,
+                            'c-disabled': option.disabled,
+                            'c-danger': option.isDangerous,
+                          })}>
                           {option.label}
                         </span>
                         {option.toggle && (
@@ -232,6 +243,9 @@ Menu.propTypes = {
     PropTypes.shape({
       label: PropTypes.node,
       onClick: PropTypes.func,
+      disabled: PropTypes.bool,
+      /** Optional testid property */
+      testId: PropTypes.string,
       /** whether option has inline toggle */
       toggle: PropTypes.shape({
         checked: PropTypes.bool,